backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
tensorflow==2.20.0
keras>=3.0.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
